\name{copynumbR-package}
\alias{copynumbR-package}
\docType{package}
\title{Useful helper functions for working with copy number data}
\description{Copy number helper functions, mainly for visualizing and working with TCGA}
\details{
\tabular{ll}{Package: \tab copynumbR\cr
Type: \tab Package\cr
Title: \tab Useful helper functions for working with copy number data\cr
Version: \tab 0.1\cr
Date: \tab 2012-03-02\cr
Author: \tab Markus Riester\cr
Maintainer: \tab Markus Riester <markus@jimmy.harvard.edu>\cr
LazyData: \tab true\cr
Imports: \tab 
CNTools,
Biobase,
ggplot2(>= 0.9.2),
TTR,
bioDist,
plyr(>= 1.8),
RColorBrewer,
scales,
grid,
xts(>= 0.9),
ggdendro,
biomaRt,
Ringo\cr
Depends: \tab 
R (>= 2.11.0),
Biobase,
CNTools,
TTR\cr
License: \tab GPL-3\cr
URL: \tab https://github.com/lima1/copynumbR\cr}
}
\author{Markus Riester}

\keyword{ package }


