\name{copynumbR.boxplot.single}
\alias{copynumbR.boxplot.single}
\title{copynumbR boxplot single}

\usage{copynumbR.boxplot.single(cn, expr, cutoffs = c(-Inf, -1.3, -0.2, 
    0.2, 0.9, Inf), cutoff.labels = c("High Loss", "Loss", "Normal", 
    "Gain", "Amplification"), min.samples = 3, gene.name, highlight = NULL, 
    highlight.labels = NULL)}
\arguments{
  \item{cn}{
}
  \item{expr}{
}
  \item{cutoffs}{
}
  \item{cutoff.labels}{
}
  \item{min.samples}{
}
  \item{gene.name}{
}
  \item{highlight}{
}
  \item{highlight.labels}{
}
}



\author{Markus Riester}





