\name{copynumbR.clone.gistic.eset}
\alias{copynumbR.clone.gistic.eset}
\title{copynumbR clone gistic eset}
\description{Extract matching copy numbers }
\usage{copynumbR.clone.gistic.eset(eset.gistic, eset.segmented)}
\arguments{
  \item{eset.gistic}{An ExpressionSet with GISTIC peaks read with copynumbR.gistic.read.lesions}
  \item{eset.segmented}{
}
}

\value{An ExpressionSet containing copy numbers from eset.segmented of the GISTIC peaks    
in eset.gistic}

\author{Markus Riester}





