\name{copynumbR.eset}
\alias{copynumbR.eset}
\title{copynumbR eset}
\description{Generate an ExpressionSet from a data.frame containing the data and a
data.frame containing clinical data}
\usage{copynumbR.eset(data, clinical, data.col = 10, id.col = 1, os.col = NULL, 
    os.event.col = NULL, dss.col = NULL, dss.event.col = NULL)}
\arguments{
  \item{data}{data.frame containing genomic data (expression, copy number, ...)}
  \item{clinical}{data.frame containing matched clinical data}
  \item{data.col}{Sample data starts at this column}
  \item{id.col}{The column of the matching sample ids in the clinical data.frame  }
  \item{os.col}{
}
  \item{os.event.col}{
}
  \item{dss.col}{
}
  \item{dss.event.col}{
}
}



\author{Markus Riester}





