\name{copynumbR.gistic.arrayfile}
\alias{copynumbR.gistic.arrayfile}
\title{copynumbR gistic arrayfile}
\description{Write GISTIC input arrayfile }
\usage{copynumbR.gistic.arrayfile(labels, file)}
\arguments{
  \item{labels}{The sample names as used in the segmented file}
  \item{file}{The output filename of this function}
}



\author{Markus Riester}





