\name{copynumbR.gistic.genes.band}
\alias{copynumbR.gistic.genes.band}
\title{copynumbR gistic genes band}
\description{Extract the GISTIC target genes }
\usage{copynumbR.gistic.genes.band(filename, band)}
\arguments{
  \item{filename}{The GISTIC output file amp_genes.conf_95.txt}
  \item{band}{The chromosome band}
}

\value{Genes in the specified chromosome band    }

\author{Markus Riester}





