\name{copynumbR.gistic.genes.region}
\alias{copynumbR.gistic.genes.region}
\title{copynumbR gistic genes region}
\description{Extract GISTIC genes in a specified region.}
\usage{copynumbR.gistic.genes.region(filename, region)}
\arguments{
  \item{filename}{The filename of the GISTIC output file amp_genes.conf_95.txt.}
  \item{region}{The region to extract. }
}

\value{Genes in the specified region.    }

\author{Markus Riester}





