\name{copynumbR.gistic.write.arrayfile}
\alias{copynumbR.gistic.write.arrayfile}
\title{copynumbR gistic write arrayfile}
\description{Write GISTIC input arrayfile. }
\usage{copynumbR.gistic.write.arrayfile(labels, file)}
\arguments{
  \item{labels}{The sample names as used in the segmented file.}
  \item{file}{The output filename of this function.}
}



\author{Markus Riester}





