\name{copynumbR.heatmap}
\alias{copynumbR.heatmap}
\title{copynumbR heatmap}

\usage{copynumbR.heatmap(eset, window = 1e+06, gain = 0.1, loss = -0.1, 
    Colv = NULL, distfun = spearman.dist, hclustfun = hclust, 
    from.chr = 1, to.chr = 22, start = NULL, end = NULL, ensembl = NULL, 
    plot = TRUE, hide.labels = FALSE, ylab = ("Chromosome"))}
\arguments{
  \item{eset}{
}
  \item{window}{
}
  \item{gain}{
}
  \item{loss}{
}
  \item{Colv}{
}
  \item{distfun}{
}
  \item{hclustfun}{
}
  \item{from.chr}{
}
  \item{to.chr}{
}
  \item{start}{
}
  \item{end}{
}
  \item{ensembl}{
}
  \item{plot}{
}
  \item{hide.labels}{
}
  \item{ylab}{
}
}



\author{Markus Riester}





