\name{copynumbR.nexus.abberations}
\alias{copynumbR.nexus.abberations}
\title{copynumbR nexus abberations}

\usage{copynumbR.nexus.abberations(filename, event, clinical, gene = FALSE, 
    data.col = ifelse(gene, 6, 4), glad.output = NULL, ...)}
\arguments{
  \item{filename}{
}
  \item{event}{
}
  \item{clinical}{
}
  \item{gene}{
}
  \item{data.col}{
}
  \item{glad.output}{
}
  \item{\dots}{
}
}



\author{Markus Riester}





