\name{copynumbR.tcga.gisticinput}
\alias{copynumbR.tcga.gisticinput}
\title{copynumbR tcga gisticinput}
\description{Create a GISTIC input file from TCGA Level 3 data}
\usage{copynumbR.tcga.gisticinput(path = ".", output.seg = "tcga.seg", 
    output.alf = "tcga.alf", hg = "hg18", verbose = TRUE)}
\arguments{
  \item{path}{The path containing the Level_3 folder}
  \item{output.seg}{The GISTIC segemented data input file}
  \item{output.alf}{The GISTIC array input file}
  \item{hg}{The genome version}
  \item{verbose}{Print some additional progress information}
}



\author{Markus Riester}





