\name{copynumbR.tcga.input}
\alias{copynumbR.tcga.input}
\title{copynumbR tcga input}

\usage{copynumbR.tcga.input(path = ".", output = "tcga.txt", verbose = TRUE, 
    grepregex = "\\\.data.txt", row.names = 1, var.column = 1)}
\arguments{
  \item{path}{
}
  \item{output}{
}
  \item{verbose}{
}
  \item{grepregex}{
}
  \item{row.names}{
}
  \item{var.column}{
}
}



\author{Markus Riester}





