\name{PMID17099711.GPL91_eset}
\alias{PMID17099711.GPL91_eset}
\docType{data}
\title{
Stransky et al. bladder microarray data
}
\description{
Provides RMA normalized and curated expression of the Stranky et al. dataset.
}
\usage{
data(PMID17099711.GPL91_eset)
}
\format{
A Bioconductor ExpressionSet containing the data.
}
\source{
From the yet unpublished curatedBladderData package.
}
\examples{
data(PMID17099711.GPL91_eset)
head(PMID17099711.GPL91_eset)
}
\keyword{datasets}
