\name{copynumbR-package}
\alias{copynumbR-package}
\docType{package}
\title{Useful helper functions for working with copy number data}
\description{Copy number helper functions, mainly for visualizing and working with TCGA}
\details{
\tabular{ll}{Package: \tab copynumbR\cr
Type: \tab Package\cr
Title: \tab Useful helper functions for working with copy number data\cr
Version: \tab 0.1\cr
Date: \tab 2012-03-02\cr
Author: \tab Markus Riester\cr
Maintainer: \tab Markus Riester <markus@jimmy.harvard.edu>\cr
Depends: \tab R (>= 2.10.0), CNTools, Biobase\cr
Suggests: \tab ggplot2, TTR, bioDist, plyr\cr
License: \tab Artistic-2.0\cr}
}
\author{Markus Riester}

\keyword{ package }


