\name{copynumbR.tcga.write.segmented}
\alias{copynumbR.tcga.write.segmented}
\title{copynumbR tcga write segmented}
\description{Create a segmented copy number file from TCGA Level 3 data}
\usage{copynumbR.tcga.write.segmented(path = ".", file = "tcga.seg", 
    hg = "hg18", verbose = TRUE)}
\arguments{
  \item{path}{The path containing the Level_3 folder.}
  \item{file}{The segemented data output file.}
  \item{hg}{The genome version.}
  \item{verbose}{Print some additional progress information.}
}

\value{A data.frame containing the segmented data.   }

\author{Markus Riester}





