\name{theme_classic2}
\alias{theme_classic2}
\title{theme classic2}
\description{A simple ggplot2 theme, no fancy stuff}
\usage{theme_classic2(base_size = 12, base_family = "")}
\arguments{
  \item{base_size}{Theme base font size}
  \item{base_family}{Theme base font family}
}



\author{Markus Riester}





